Add sonatype snapshot repository to sbt
